StreamExtended moved to TWP project
