--- src/StreamExtended/StreamExtended.csproj
@@ -1,21 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>net45;netstandard2.0</TargetFrameworks>
-    <SignAssembly>True</SignAssembly>
-    <AssemblyOriginatorKeyFile>StrongKey.snk</AssemblyOriginatorKeyFile>
-    <Version>1.0.0</Version>
-    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-    <Description>Package Description</Description>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Remove="Network\ClientHelloAlpnAdderStream.cs" />
-    <Compile Remove="Network\ServerHelloAlpnAdderStream.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Remove="StreamExtended.nuspec" />
-  </ItemGroup>
-
-</Project>
