+++ src/Titanium.Web.Proxy/EventArguments/SessionEventArgsBase.cs
@@ -2,13 +2,13 @@
 using System.Collections.Generic;
 using System.Net;
 using System.Threading;
-using StreamExtended;
-using StreamExtended.Network;
 using Titanium.Web.Proxy.Helpers;
 using Titanium.Web.Proxy.Http;
 using Titanium.Web.Proxy.Models;
 using Titanium.Web.Proxy.Network;
 using Titanium.Web.Proxy.Network.Tcp;
+using Titanium.Web.Proxy.StreamExtended.BufferPool;
+using Titanium.Web.Proxy.StreamExtended.Network;
 
 namespace Titanium.Web.Proxy.EventArguments
 {
