+++ src/Titanium.Web.Proxy/Extensions/SslExtensions.cs
@@ -5,7 +5,7 @@ using System.Security.Authentication;
 using System.Security.Cryptography.X509Certificates;
 using System.Threading;
 using System.Threading.Tasks;
-using StreamExtended;
+using Titanium.Web.Proxy.StreamExtended;
 
 namespace Titanium.Web.Proxy.Extensions
 {
