+++ src/Titanium.Web.Proxy/Extensions/StreamExtensions.cs
@@ -2,7 +2,7 @@
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
-using StreamExtended;
+using Titanium.Web.Proxy.StreamExtended.BufferPool;
 
 namespace Titanium.Web.Proxy.Extensions
 {
