+++ src/Titanium.Web.Proxy/Helpers/HttpResponseWriter.cs
@@ -2,8 +2,8 @@
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
-using StreamExtended;
 using Titanium.Web.Proxy.Http;
+using Titanium.Web.Proxy.StreamExtended.BufferPool;
 
 namespace Titanium.Web.Proxy.Helpers
 {
