+++ src/Titanium.Web.Proxy/StreamExtended/Network/CopyStream.cs
@@ -1,8 +1,9 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Titanium.Web.Proxy.StreamExtended.BufferPool;
 
-namespace StreamExtended.Network
+namespace Titanium.Web.Proxy.StreamExtended.Network
 {
     /// <summary>
     ///     Copies the source stream to destination stream.
