+++ src/Titanium.Web.Proxy/StreamExtended/Network/CustomBufferedPeekStream.cs
@@ -1,7 +1,8 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
+using Titanium.Web.Proxy.StreamExtended.BufferPool;
 
-namespace StreamExtended.Network
+namespace Titanium.Web.Proxy.StreamExtended.Network
 {
     internal class CustomBufferedPeekStream : ICustomStreamReader
     {
