+++ src/Titanium.Web.Proxy/StreamExtended/ServerHelloInfo.cs
@@ -1,15 +1,15 @@
-using StreamExtended.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using Titanium.Web.Proxy.StreamExtended.Models;
 
-namespace StreamExtended
+namespace Titanium.Web.Proxy.StreamExtended
 {
     /// <summary>
-    /// Wraps up the client SSL hello information.
+    /// Wraps up the server SSL hello information.
     /// </summary>
-    public class ClientHelloInfo
+    public class ServerHelloInfo
     {
         private static readonly string[] compressions = {
             "null",
@@ -41,11 +41,11 @@ namespace StreamExtended
 
         public byte[] SessionId { get; set; }
 
-        public int[] Ciphers { get; set; }
+        public int CipherSuite { get; set; }
 
-        public byte[] CompressionData { get; set; }
+        public byte CompressionMethod { get; set; }
 
-        internal int ClientHelloLength { get; set; }
+        internal int ServerHelloLength { get; set; }
 
         internal int EntensionsStartPosition { get; set; }
 
@@ -77,7 +77,7 @@ namespace StreamExtended
         public override string ToString()
         {
             var sb = new StringBuilder();
-            sb.AppendLine($"A SSLv{HandshakeVersion}-compatible ClientHello handshake was found. Titanium extracted the parameters below.");
+            sb.AppendLine($"A SSLv{HandshakeVersion}-compatible ServerHello handshake was found. Titanium extracted the parameters below.");
             sb.AppendLine();
             sb.AppendLine($"Version: {SslVersionToString(MajorVersion, MinorVersion)}");
             sb.AppendLine($"Random: {string.Join(" ", Random.Select(x => x.ToString("X2")))}");
@@ -93,28 +93,18 @@ namespace StreamExtended
                 }
             }
 
-            if (CompressionData != null && CompressionData.Length > 0)
+            string compression = compressions.Length > CompressionMethod 
+                ? compressions[CompressionMethod] 
+                : $"unknown [0x{CompressionMethod:X2}]";
+            sb.AppendLine($"Compression: {compression}");
+
+            sb.Append("Cipher:");
+            if (!SslCiphers.Ciphers.TryGetValue(CipherSuite, out string cipherStr))
             {
-                int compressionMethod = CompressionData[0];
-                string compression = compressions.Length > compressionMethod 
-                    ? compressions[compressionMethod] 
-                    : $"unknown [0x{compressionMethod:X2}]";
-                sb.AppendLine($"Compression: {compression}");
+                cipherStr = "unknown";
             }
 
-            if (Ciphers.Length > 0)
-            {
-                sb.AppendLine("Ciphers:");
-                foreach (int cipherSuite in Ciphers)
-                {
-                    if (!SslCiphers.Ciphers.TryGetValue(cipherSuite, out string cipherStr))
-                    {
-                        cipherStr = "unknown";
-                    }
-
-                    sb.AppendLine($"[0x{cipherSuite:X4}] {cipherStr}");
-                }
-            }
+            sb.AppendLine($"[0x{CipherSuite:X4}] {cipherStr}");
 
             return sb.ToString();
         }
