+++ src/Titanium.Web.Proxy/StreamExtended/SslTools.cs
@@ -1,11 +1,11 @@
-﻿using StreamExtended.Models;
-using StreamExtended.Network;
-using System.Collections.Generic;
-using System.IO;
+﻿using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using Titanium.Web.Proxy.StreamExtended.BufferPool;
+using Titanium.Web.Proxy.StreamExtended.Models;
+using Titanium.Web.Proxy.StreamExtended.Network;
 
-namespace StreamExtended
+namespace Titanium.Web.Proxy.StreamExtended
 {
     /// <summary>
     /// Use this class to peek SSL client/server hello information.
