+++ src/Titanium.Web.Proxy/Titanium.Web.Proxy.Mono.csproj
@@ -16,10 +16,6 @@
     <PackageReference Include="Portable.BouncyCastle" Version="1.8.5" />
   </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\StreamExtended\StreamExtended.csproj" />
-  </ItemGroup>
-
   <ItemGroup Condition="'$(TargetFramework)' == 'net45'">
     <Reference Include="System.Web" />
   </ItemGroup>
