--- src/Titanium.Web.Proxy/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="BouncyCastle" version="1.8.3.1" targetFramework="net45" />
-  <package id="BrotliSharpLib" version="0.3.1" targetFramework="net45" />
-  <package id="StreamExtended" version="1.0.190" targetFramework="net45" />
-</packages>
